+++ packages/Webkul/Razorpay/src/Models/RazorpayEventProxy.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Webkul\Razorpay\Models;
+
+use Konekt\Concord\Proxies\ModelProxy;
+
+class RazorpayEventProxy extends ModelProxy {}