+++ packages/Webkul/Razorpay/src/Payment/RazorpayPayment.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Webkul\Razorpay\Payment;
+
+use Illuminate\Support\Facades\Storage;
+use Webkul\Payment\Payment\Payment;
+
+class RazorpayPayment extends Payment
+{
+    /**
+     * Payment method code.
+     *
+     * @var string
+     */
+    protected $code = 'razorpay';
+
+    /**
+     * Get redirect url.
+     */
+    public function getRedirectUrl()
+    {
+        return route('razorpay.payment.redirect');
+    }
+
+    /**
+     * Returns payment method image.
+     *
+     * @return array
+     */
+    public function getImage()
+    {
+        $url = $this->getConfigData('image');
+
+        return $url ? Storage::url($url) : bagisto_asset('images/razorpay.png', 'shop');
+    }
+}