+++ packages/Webkul/Razorpay/src/Resources/lang/ar/app.php
@@ -1,0 +1,16 @@
+<?php
+
+return [
+    'drop-in-ui' => [
+        'title' => 'رايزورباي',
+    ],
+
+    'response' => [
+        'payment' => [
+            'cancelled' => 'تم إلغاء دفعة Razorpay.',
+        ],
+
+        'something-went-wrong'     => 'حدث خطأ ما.',
+        'supported-currency-error' => 'العملة :currency غير مدعومة. العملات المدعومة هي: :supportedCurrencies.',
+    ],
+];