+++ packages/Webkul/Razorpay/src/Resources/lang/id/app.php
@@ -1,0 +1,16 @@
+<?php
+
+return [
+    'drop-in-ui' => [
+        'title' => 'Razorpay',
+    ],
+
+    'response' => [
+        'payment' => [
+            'cancelled' => 'Pembayaran Razorpay telah dibatalkan.',
+        ],
+
+        'something-went-wrong'     => 'Ada yang salah.',
+        'supported-currency-error' => 'Mata uang :currency tidak didukung. Mata uang yang didukung: :supportedCurrencies.',
+    ],
+];