+++ packages/Webkul/Razorpay/src/Resources/lang/it/app.php
@@ -1,0 +1,16 @@
+<?php
+
+return [
+    'drop-in-ui' => [
+        'title' => 'Razorpay',
+    ],
+
+    'response' => [
+        'payment' => [
+            'cancelled' => 'Il pagamento Razorpay è stato annullato.',
+        ],
+
+        'something-went-wrong'     => 'Qualcosa è andato storto.',
+        'supported-currency-error' => 'La valuta :currency non è supportata. Valute supportate: :supportedCurrencies.',
+    ],
+];