+++ packages/Webkul/Razorpay/src/Resources/lang/ja/app.php
@@ -1,0 +1,16 @@
+<?php
+
+return [
+    'drop-in-ui' => [
+        'title' => 'Razorpay',
+    ],
+
+    'response' => [
+        'payment' => [
+            'cancelled' => 'Razorpayの支払いがキャンセルされました。',
+        ],
+
+        'something-went-wrong'     => '問題が発生しました。',
+        'supported-currency-error' => '通貨 :currency はサポートされていません。サポートされている通貨: :supportedCurrencies.',
+    ],
+];